Ai speed up when player is looking at certain objects
I wrote a simple script that can be attached to any object. It makes the ai speed up when the object is in view. Said object must be rendered for the code to work, so it can't be an empty marker. If more than one objects are at view, and they try to set the speed to different values, maximum value will be used.

Also unity keeps creating these .meta files so I'm including them as well.